+++ tests/deprecated_api/test_remove_1-8.py
@@ -106,3 +106,8 @@ def test_v1_8_0_deprecated_on_hpc_hooks(tmpdir):
         match=r"Method `LightningModule.on_hpc_load` is deprecated in v1.6 and will be removed in v1.8."
     ):
         trainer.fit(load_model)
+
+def test_v1_8_0_deprecated_trainer_should_rank_save_checkpoint(tmpdir):
+    trainer = Trainer()
+    with pytest.deprecated_call(match=r"`Trainer.should_rank_save_checkpoint` is deprecated in v1.6 and will be removed in 1.8."):
+        _ = trainer.should_rank_save_checkpoint